Add files via upload
Underwater mines are common explosive bombs that are placed around a coastal region of a country, deep under the ocean to ensure safety, defence and security. This could be a threat to a submarine or any civilian boats that crosses over the mine. Though it provides enormous security, it is a threat to marines which may mistake it as a rock. Due to its similar appearance, size and placings in ocean bed, a rock can appear to be and mine and vice versa. Moreover, absence of light under ocean due to attenuation makes it impossible to capture a image to be processed. Here, SONAR technique is used to explore the ocean bed and identify rocks and mines accurately. The dataset contains 60 different angles under which rocks and mines are recorded with specific value. The main objective is to enhance the model using Ensemble learning techniques. Standalone algorithms to overcome overfitting, under fitting and feature selection. Bagging to improve stability and reduce bias value. Along with frequency values, underwater sonar images are processed to determine the object detection to be rock or mine. This focuses on providing double standards on the object that has been encountered. When frequency waves are processed through one model, other model processes the sonar image captured. Ensemble techniques has been performed on the sonar values. It is observed that KNN fits the best before and after scaling the algorithm. RandomForest classifier has been used for bagging the model that reduces model variance and improves model stability. It combines multiple decisions made by the model and combines the previous predictions to make a robust system and to make accurate predictions. Boosting method such as AdaBoost has been employed to correct the errors made by the previous model and iteratively adjust the wights accordingly. It eliminates the weak learners and out of point learners to emphasize a good boosting in the model. Thus, to boost the overall ensemble technique, bagging and boosting method has been inculcated. As to classify the sonar images, convolutional network is used. It works the best for images classification to scan the input images using kernels to detect the image points and create a feature map. To not encounter such disaster, it is crucial to differentiate them and warn the submarines accordingly.